requirements.txt
PyQt5==5.15.9
pandas==2.0.3
openpyxl==3.1.2
pywin32==306
xlrd==2.0.1
pillow==10.0.0
pyinstaller==6.0.0 